requirements.txt
asgiref==3.6.0
attrs==21.2.0
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
colorama==0.4.6
cryptography==3.4.8
distro==1.7.0
Django==5.0.1
django-ckeditor==6.7.0
django-js-asset==2.2.0
filelock==3.6.0
future==0.18.3
idna==2.10
Jinja2==3.0.3
MarkupSafe==2.1.4
more-itertools==10.2.0
numpy==1.21.6
olefile==0.47
packaging==21.3
pillow==10.2.0
platformdirs==2.5.4
psycopg==3.1.17
psycopg2-binary==2.9.9
pycparser==2.21
pyparsing==3.1.1
PyYAML==6.0.1
requests==2.25.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.18
zipp==1.0.0
